разделение на компонеты
